requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.3
attrs==23.1.0
boto3==1.26.118
botocore==1.29.123
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-storages==1.13.2
django-tinymce==3.6.1
exceptiongroup==1.1.2
factory-boy==3.3.0
Faker==19.3.0
frozenlist==1.4.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
jmespath==1.0.1
multidict==6.0.4
openai==0.28.1
packaging==23.1
Pillow==9.5.0
pluggy==1.2.0
psycopg2-binary==2.9.6
pytest==7.4.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
python-dateutil==2.8.2
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==1.26.15
yarl==1.9.2


